DT-185 moving tools to a separate file
