+++ src/components/filter/Filter.module.scss
@@ -10,7 +10,6 @@
   justify-content: flex-start;
   padding: 5px;
   font-size: 1.7rem;
-  ;
   color: (--primary-text-color);
 }
