+++ src/logic/catalogPage/CatalogPage.module.scss
